BE/requirements.txt
fastapi==0.112.1
starlette==0.38.2
uvicorn==0.30.6
httpx==0.27.0
numpy==1.26.4
scipy==1.14.1
pandas==2.2.2
scikit-learn==1.5.1
torch==2.4.0
transformers==4.44.2
sentence-transformers==3.0.1
huggingface-hub==0.24.6
tokenizers==0.19.1
faiss-cpu==1.8.0.post1
pinecone-client==5.0.1
pinecone-plugin-inference==1.0.3
pinecone-plugin-interface==0.0.7
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2
annotated-types==0.7.0
typing_extensions==4.12.2
pydantic_core==2.20.1
rich==13.7.1
